Bootcamps/Microsoft Certification Challenge #5 - AI 102/Tradutor-de-Artigos-Tecnicos-com-AzureAI/requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
python-dotenv==1.2.1
azure-ai-translation-text==1.0.1
